# Pass parameter vectors by reference