--- a/include/GRANSAC.hpp
+++ b/include/GRANSAC.hpp
@@ -33,9 +33,9 @@ namespace GRANSAC
 
     public:
 	RANSAC(void)
-	{
+        {
 	    int nThreads = std::max(1, omp_get_max_threads());
-	    std::cout << "[ INFO ]: Maximum usable threads: " << nThreads << std::endl;
+            std::cout << "[ INFO ]: Maximum usable threads: " << nThreads << std::endl;
 	    for(int i = 0; i < nThreads; ++i)
 	    {
 		std::random_device SeedDevice;
@@ -66,11 +66,11 @@ namespace GRANSAC
 	std::shared_ptr<T> GetBestModel(void) { return m_BestModel; };
 	const std::vector<std::shared_ptr<AbstractParameter>>& GetBestInliers(void) { return m_BestInliers; };
 
-	bool Estimate(std::vector<std::shared_ptr<AbstractParameter>> Data)
+        bool Estimate(const std::vector<std::shared_ptr<AbstractParameter>> &Data)
 	{
 	    if(Data.size() <= t_NumParams)
 	    {
-		std::cout << "[ WARN ]: RANSAC - Number of data points is too less. Not doing anything." << std::endl;
+                std::cout << "[ WARN ]: RANSAC - Number of data points is too less. Not doing anything." << std::endl;
 		return false;
 	    }
 
@@ -118,7 +118,7 @@ namespace GRANSAC
 		}
 	    }
 
-	    // std::cout << "BestInlierFraction: " << m_BestModelScore << std::endl;
+            // std::cout << "BestInlierFraction: " << m_BestModelScore << std::endl;
 
 	    Reset();