Hi! Thanks for the library, I found it quite useful when working on my Bachelor thesis. I had to intersect >2 noisy lines in a single point, and GRANSAC provided an easy way to test whether a RANSAC-like solution works in my problem (and yes, it does).

However, I noticed a significant performance drop when using RANSAC. I'm trying to fit a large number of objects (up to 4000) on every frame of a 30fps video. Upon profiling the program, I found out that we can improve by passing the parameters are passed by reference rather than by value, so that it doesn't copy 4000 shared_ptr's every time. It helped a bit, and I think there's no good reason for copying those vectors anyway. This PR in a nutshell:

```
void GRANSAC::SomeMethod(vector<AbstractParameterPtr> params)        // old
void GRANSAC::SomeMethod(const vector<AbstractParameterPtr> &params) // new
```

I also made the library to output messages to stderr instead of stdout. Stdout is really for the program's output data, and library warnings do not belong there.

Cheers!