@drsrinathsridhar 2017-06-22
Hi!

Glad to see the library of use to you. Are you using OpenMP for speedup?

Thanks also for the pull request. I agree that passing by reference is a good idea.
But before I can merge your changes, did you unit test them on the line fitting example?

About stdout/stderr: Honestly, I think the "proper" way to do it is to have a custom stream to which all messages are written to which the API caller can send out in batch to stdout/stderr if required.

Besides, stderr is only for error/warning messages. GRANSAC also shows information (> stdout) in addition to warnings and errors (> stderr).