@Norrius 2017-06-22
Since I have OpenCV 3.1, building the example requires some changes. Namely, including <opencv2/imgproc.hpp> and removing the version requirement from CMake config. I don't have OpenCV 2 so I can't check if it has this header (you could just add it to support OpenCV 3 if it doesn't break anything).

With this in mind, the line fitting example works fine on my system. I also tried running the example 1000 times with and without the patch to see if there's a noticeable change in performance.
Without: 15.74 ms avg, 2.64 ms stddev
With: 13.36 ms avg, 2.17 ms stddev

While the result is probably not statistically significant, it's nice to see that at the very least it doesn't lead to regression.

Well, I mentioned the stdout/stderr thing because I was trying to analyse my program's output in another program (written in Python), and I've chosen to pipe the data through stdout. At some random point there were too few objects to fit, the library produced a warning, and the parser broke. I won't insist, but I still think stderr makes more sense here.