@drsrinathsridhar 2017-06-23
Thanks, I removed the dependency on OpenCV 2.4. It now works with any OpenCV version.

I merged your pull request with the exception of changing the INFO to stdout. The WARN will redirect to stderr as you suggested.